requirements.txt
fastapi
scenedetect
opencv-python
python-dotenv
PyGithub
uvicorn
python-multipart
langchain
google-generativeai
langchain-google-genai
